refactor: CardMapper.kt 삭제 및 CardUiModel.kt로 통합
