+++ app/src/main/java/woowacourse/payments/ui/CardUiModel.kt
@@ -3,8 +3,6 @@ package woowacourse.payments.ui
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 import woowacourse.payments.domain.Card
-import woowacourse.payments.ui.util.formatCardNumber
-import woowacourse.payments.ui.util.formatExpired
 
 @Parcelize
 data class CardUiModel(
@@ -15,6 +13,20 @@ data class CardUiModel(
     val maskedNumber get() = formatCardNumber(number)
     val formattedExpired get() = formatExpired(expired)
 }
+
+fun formatCardNumber(cardNumber: String): String {
+    val front = cardNumber.take(8).chunked(4).joinToString(" - ")
+
+    return if (cardNumber.length <= 8) {
+        front
+    } else {
+        val maskedCount = (cardNumber.length - 8).coerceAtLeast(0)
+        val masked = "*".repeat(maskedCount).chunked(4).joinToString(" - ")
+        "$front - $masked"
+    }
+}
+
+fun formatExpired(expired: String): String = expired.chunked(2).joinToString(" / ")
 
 fun Card.toPresentation(): CardUiModel =
     CardUiModel(
