--- app/src/main/java/woowacourse/payments/ui/util/CardMapper.kt
@@ -1,15 +1,0 @@
-package woowacourse.payments.ui.util
-
-fun formatCardNumber(cardNumber: String): String {
-    val front = cardNumber.take(8).chunked(4).joinToString(" - ")
-
-    return if (cardNumber.length <= 8) {
-        front
-    } else {
-        val maskedCount = (cardNumber.length - 8).coerceAtLeast(0)
-        val masked = "*".repeat(maskedCount).chunked(4).joinToString(" - ")
-        "$front - $masked"
-    }
-}
-
-fun formatExpired(expired: String): String = expired.chunked(2).joinToString(" / ")
